// corexverse/src/generated/models/Mdi_project_listsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Mdi_project_listsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Mdi_project_listsstatecode = keyof typeof Mdi_project_listsstatecode;
export const Mdi_project_listsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Mdi_project_listsstatuscode = keyof typeof Mdi_project_listsstatuscode;

export interface Mdi_project_lists {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  mdi_cg?: string;
  mdi_cgname?: string;
  mdi_customer?: string;
  mdi_customername?: string;
  mdi_golivedate?: string;
  mdi_id: string;
  mdi_maintenance_period_years?: string;
  mdi_mainworklocation?: string;
  mdi_mainworklocationname?: string;
  mdi_name?: string;
  mdi_offshore?: string;
  mdi_offshorename?: string;
  mdi_oppid?: string;
  mdi_project_description?: string;
  mdi_project_listid: string;
  mdi_project_status?: string;
  mdi_project_statusname?: string;
  mdi_support_fte?: string;
  mdi_wbscode?: string;
  mdi_workload_mm?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Mdi_project_listsstatecode;
  statecodename?: string;
  statuscode?: Mdi_project_listsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}